==> 02-data.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Data sources
```{r}
library(rvest)
library(readr)
library(quantmod)
```

## Quarterly Revenue of tech companies
The data is collected by using rvest package to scrap data from macrotrends.net. The code below is an example of how data is scraped using Netflix. Other data gathering methods considered are grabbing data directly from each company's financial report as well as copy and paste data from Yahoo Finance. However, I choose to scrap the data because it is the easiest way to download data for multiple companies and all the data has been validated with the company's public financial report or Yahoo Finance to prevent errors. Some issue with the data now is that both variables are in character type and we would like the first to be date variable and second to be numerical. This problem will be handled in Data transformation part. 
```{r}
netflix = read_html('https://www.macrotrends.net/stocks/charts/NFLX/netflix/revenue')
disney = read_html('https://www.macrotrends.net/stocks/charts/DIS/disney/revenue')
amazon = read_html('https://www.macrotrends.net/stocks/charts/AMZN/amazon/revenue')
att = read_html('https://www.macrotrends.net/stocks/charts/T/at-t/revenue')
roku = read_html('https://www.macrotrends.net/stocks/charts/ROKU/roku/revenue')
overstock = read_html('https://www.macrotrends.net/stocks/charts/OSTK/overstock/revenue')
blue_apron = read_html('https://www.macrotrends.net/stocks/charts/APRN/blue-apron-holdings/revenue')
wayfair = read_html('https://www.macrotrends.net/stocks/charts/W/wayfair/revenue')
etsy = read_html('https://www.macrotrends.net/stocks/charts/ETSY/etsy/revenue')
chewy = read_html('https://www.macrotrends.net/stocks/charts/CHWY/chewy/revenue')
farfetch = read_html('https://www.macrotrends.net/stocks/charts/FTCH/farfetch/revenue')
ea = read_html('https://www.macrotrends.net/stocks/charts/EA/electronic-arts/revenue')
sciplay = read_html('https://www.macrotrends.net/stocks/charts/SCPL/sciplay/revenue')
ttwo = read_html('https://www.macrotrends.net/stocks/charts/TTWO/take-two-interactive-software/revenue')
atvi = read_html('https://www.macrotrends.net/stocks/charts/ATVI/activision-blizzard/revenue')
zoom = read_html('https://www.macrotrends.net/stocks/charts/ZM/zoom-video-communications/revenue')
csco = read_html('https://www.macrotrends.net/stocks/charts/CSCO/cisco/revenue')
eght = read_html('https://www.macrotrends.net/stocks/charts/EGHT/8x8-inc/revenue')
rng = read_html('https://www.macrotrends.net/stocks/charts/RNG/ringcentral/revenue')
msft = read_html('https://www.macrotrends.net/stocks/charts/MSFT/microsoft/revenue')
goog = read_html('https://www.macrotrends.net/stocks/charts/GOOG/alphabet/revenue')
twtr = read_html('https://www.macrotrends.net/stocks/charts/TWTR/twitter/revenue')
fb = read_html('https://www.macrotrends.net/stocks/charts/FB/facebook/revenue')
aapl = read_html('https://www.macrotrends.net/stocks/charts/AAPL/apple/revenue')

companies = list(netflix, disney, amazon, att, roku, overstock, blue_apron, wayfair, etsy, chewy, farfetch, ea, sciplay, ttwo, 
           atvi, zoom, csco, eght, rng, msft, goog, twtr, fb, aapl)
company_names = c('netflix', 'disney', 'amazon', 'att', 'roku', 'overstock', 'blue_apron', 'wayfair', 'etsy', 'chewy', 'farfetch', 'ea', 'sciplay', 'ttwo', 'atvi', 'zoom', 'csco', 'eght', 'rng', 'msft', 'goog', 'twtr', 'fb', 'aapl')

df_list = list()
for (i in 1:24) {
  q = companies[[i]] %>% 
  html_nodes(xpath = '//*[@id="style-1"]/div[2]/table') %>% html_table
  
  col_names <- c('Date', company_names[i])
  q = data.frame(q)[1:8, ]
  names(q) <- col_names
  q[, company_names[i]] <- as.numeric(gsub("[\\$,]", "", q[,2]))
  df_list[[company_names[i]]] = q

}

joined = df_list[[1]]
for (j in 2:24) {
  joined = merge(joined, df_list[[j]], by = 'Date', all.x = TRUE)
}

joined = joined[-c(11, 17, 18)]
write.csv(joined, file='./data/quarterly_rev.csv')
```

### Example for extracting Netflix quarterly revenue 
```{r}
test = read_html('https://www.macrotrends.net/stocks/charts/NFLX/netflix/revenue')
netflix = read_html('https://www.macrotrends.net/stocks/charts/NFLX/netflix/revenue')

netflix = netflix %>% 
  html_nodes(xpath = '//*[@id="style-1"]/div[2]/table') %>% html_table

netflix = data.frame(netflix)
col_names <- c('Date','Revenue')
names(netflix) = col_names
head(netflix, 10)
```


## Investors' confidence index
Data collection for individual confidence index is supported by Andrew Redleaf of Lynne and Andrew Redleaf Foundation(previously Whitebox advisors). There are two samples that include in the data:  a sample of wealthy individual investors, and a sample of institutional investors. In the years of our interests, the samples of both individual and institutional investors are purchased from InfoUSA( Data Axle: a reputable provider of data).
\n The datasets have the U.S. institutional and individual index values from 12/2018 to 10/2020. The One-Year Confidence index is the percentage of respondents giving a number strictly greater than 0 for 'in one year'.Index values are in numeric forms ranging from 0 to 100, and they are reported monthly.
\n The potential bias of data comes from two parts. Firstly, the surveys which are randomly sampled to generate datasets conducted at six-month intervals.  Thus, the index value of current month is the average of monthly surveys in the past six months.  It can be observed from the dataset that the deviation of index values among months is small due to the averaging effect, and it may cause bias when we explore the relationship of stock market and confidence index values.     Moreover, the question in the survey is worded to mention the possibilities that the respondent could predict a downturn.  Participants tend to have more passive responses than those who receive more optimistically worded questions. Thus, confidence index values in general is a subjective metric.  https://som.yale.edu/faculty-research-centers/centers-initiatives/international-center-for-finance/data/stock-market-confidence-indices/united-states-stock-market-confidence-indices

```{r}
CI <- read.csv("./data/CI.csv")
```

## Daily stock price 
For each of the companies we choosed, the historical stock price data can be quickly loaded using getSymbols function in 'quantmod' package. Using 'yahoo' as the sourcing method, we load the past daily stock prices including “Open price”, “High price”, “Low price”, “Close price”, “Volume” and “Adjusted price” from January 1, 2020 to December 11, 2020.
