==> 07-conclusion.Rmd <==
# Conclusion

This report explores several aspects of tech companies performances during COVID-19 period. By looking at the three metrics introduced in the project, we have found that different sub-industries in technology behave distinctly. We think this is due to the nature of each sub-industry. For example, we think video game industry and online streaming are born to be internet companies and their performances soar during the global pandemic.

\n On the other hand, we believe that emerging sectors such as online shopping and work-from-home that are still in the transition from traditional industry to technology industry. Currently, these sectors show a winner-take-all pattern as all companies provide interchangeable services. By our analysis based on E-commerce and video-conference companies, we believe that even though the COVID-19 created an unprecedented opportunity for them, these companies need to re-think about their diverse sources of growth to better catch the chance.  

\n All in all, by looking at the integrated companies and the general market trends provided by the Confidence Index, we consider that most of the tech companies out-performed the the whole market during COVID-19 despite the global recession. 

\n In the future, we would like to keep tracking how technology companies behave in the long run after the pandemic crash. Due to the limited data we collected, the trends we described in this project may be biased. However, we hope to provide some directions for people who want to study this topic further in the future. As what people have been saying all year round, "we lived through a time that will be in future history books."
